完善企业图片管理功能／修复系统管理员端增加省级管理员功能

+++ controllers/CompanyPictureController.php
@@ -14,6 +14,24 @@ use yii\web\UploadedFile;
 
 class CompanyPictureController extends \yii\web\Controller
 {
+    
+    public function behaviors()
+    {
+        return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['index', 'pictures-list', 'upload-picture', 'delete-picture', 'update-comment',],
+                'rules' => [
+                     [
+                        'allow' => 'true',
+                        'actions' => ['index', 'pictures-list', 'upload-picture', 'delete-picture', 'update-comment',s],
+                        'roles' => ['company'],
+                    ],
+                ]
+            ],
+        ];
+    }
+    
     /*
      * 全部图片类型的概览
      */
@@ -73,15 +91,46 @@ class CompanyPictureController extends \yii\web\Controller
         ]);
     }
     
+    /*
+     * 删除file id为$id的文件
+     * 返回删除文件所属类型的目录列表
+     */
     public function actionDeletePicture($id)
     {
-        //imageFile表通过外键级联删除
-        $file = File::find()->where(['file_id' => $id])->one();
-        $typeid = $file->ImageFile->image_typeid;
+        //imageFile表通过外键级联删除,且保证只能删除自己公司的图片
+        $file = File::find()->innerJoinWith('imageFile')->where(['yii_file.file_id' => $id, 'company_id' => $this->getCompanyId()])->one();
+        if($file == NULL){
+            return $this->redirect(['/company-picture/index']);
+        }
+        $typeid = $file->imageFile->image_typeid;
         if($file->delete()){
-            Yii::$app->session->setFlash('success', '文件删除成功！');
+            Yii::$app->session->setFlash('delete-success', '文件删除成功！');
         }
         return $this->redirect(['/company-picture/pictures-list', 'id' => $typeid]);
+    }
+    
+    /*
+     * 更新file_id为$id的b备注
+     */
+    public function actionUpdateComment($id)
+    {
+        $file = File::find()->innerJoinWith('imageFile')->where(['yii_file.file_id' => $id, 'company_id' => $this->getCompanyId()])->one();
+        if($file == NULL){
+            return $this->redirect(['/company-picture/index']);
+        }
+          if($file->load(Yii::$app->request->post())){
+              $isVaild = $file->validate();
+              if($isVaild){
+                  $imageFile = ImageFile::find()->innerJoinWith('file')->where(['yii_file.file_id' => $id, 'company_id' => $this->getCompanyId()])->one();
+                  $imageFile->image_comment = $file->file_comment;
+                  $file->save();
+                  $imageFile->save();
+                  return $this->redirect(['/company-picture/pictures-list', 'id' => $imageFile->image_typeid]);
+              }
+          }
+          return $this->render('update-comment', [
+              'model' => $file,
+          ]);
     }
